VenuesSection.tsx: inline ad-hoc VenueNames component
