--- web/components/VenueNames/index.ts
@@ -1,1 +1,0 @@
-export { VenueNames as default } from './VenueNames';
